+++ src/integrationTest/groovy/mpern/sap/commerce/ccv2/validation/UseConfigValidatorSpec.groovy
@@ -17,37 +17,73 @@ class UseConfigValidatorSpec extends Specification {
     UseConfigValidator validator
 
     def setup() {
-        validator = new UseConfigValidator(testProjectDir.root.toPath());
+        validator = new UseConfigValidator(testProjectDir.root.toPath())
     }
 
     def "locations are validated"() {
         given:
-        def rawManifest = new JsonSlurper().parse(this.getClass().getResource('/validator/useconfig-invalid-paths-manifest.json')) as Map<String, Object>
+        def rawManifest = new JsonSlurper().parseText('''\
+        {
+          "commerceSuiteVersion": "2005",
+          "useConfig": {
+            "properties": [
+              {
+                "location": "/hybris/config/environments/local-dev.properties",
+                "persona": "development"
+              },
+              {
+                "location": "../hybris/config/environments/local-stage.properties",
+                "persona": "staging"
+              },
+              {
+                "location": "./hybris/config/environments/local-prod.properties",
+                "persona": "production"
+              },
+              {
+                "location": "hybris/config/environments/common.properties"
+              },
+              {
+                "location": "hybriß/cönfig/environments/common.properties"
+              }
+            ]
+          }
+        }
+        
+        ''') as Map<String, Object>
         def manifest = Manifest.fromMap(rawManifest)
 
         when:
-        List<Error> errors = validator.validate(manifest)
+        def errors = validator.validate(manifest)
 
         then:
         errors.size() == 5
         errors.any{ it.location == "useConfig.properties[0]" && it.message.contains("absolute") }
         errors.any{ it.location == "useConfig.properties[1]" && it.message.contains("relative") }
         errors.any{ it.location == "useConfig.properties[2]" && it.message.contains("relative") }
-        errors.any{ it.location == "useConfig.properties[3]" && it.message.contains("found") }
+        errors.any{ it.location == "useConfig.properties[3]" && it.message.contains("not found") }
         errors.any{ it.location == "useConfig.properties[4]" && it.message.contains("invalid") }
     }
 
     def "localextensions.xml must be a valid localextensions.xml file"() {
         given:
-        def rawManifest = new JsonSlurper().parse(this.getClass().getResource('/validator/useconfig-localextensions-manifest.json')) as Map<String, Object>
-        Manifest m = Manifest.fromMap(rawManifest)
+        def rawManifest = new JsonSlurper().parseText('''\
+        {
+          "commerceSuiteVersion": "2005",
+          "useConfig": {
+            "extensions": {
+              "location": "localextensions.xml"
+            }
+          }
+        }
+        ''') as Map<String, Object>
+        def manifest = Manifest.fromMap(rawManifest)
 
         testProjectDir.newFile("localextensions.xml").text = """\
         <something><invalid></invalid></something>
         """.stripIndent()
 
         when:
-        List<Error> errors = validator.validate(m)
+        def errors = validator.validate(manifest)
 
         then:
         errors.size() == 1
@@ -56,8 +92,17 @@ class UseConfigValidatorSpec extends Specification {
 
     def "localextensions.xml must only contain extension names"() {
         given:
-        def rawManifest = new JsonSlurper().parse(this.getClass().getResource('/validator/useconfig-localextensions-manifest.json')) as Map<String, Object>
-        Manifest m = Manifest.fromMap(rawManifest)
+        def rawManifest = new JsonSlurper().parseText('''
+        {
+          "commerceSuiteVersion": "2005",
+          "useConfig": {
+            "extensions": {
+              "location": "localextensions.xml"
+            }
+          }
+        }
+        ''') as Map<String, Object>
+        Manifest manifest = Manifest.fromMap(rawManifest)
 
         testProjectDir.newFile("localextensions.xml").text = '''\
         <?xml version="1.0" encoding="UTF-8" standalone="no"?>
@@ -71,7 +116,7 @@ class UseConfigValidatorSpec extends Specification {
         '''.stripIndent()
 
         when:
-        List<Error> errors = validator.validate(m)
+        def errors = validator.validate(manifest)
 
         then:
         errors.size() == 1
@@ -80,12 +125,37 @@ class UseConfigValidatorSpec extends Specification {
 
     def "properties must use valid personas and aspects"() {
         given:
-        def rawManifest = new JsonSlurper().parse(this.getClass().getResource('/validator/useconfig-properties-manifest.json')) as Map<String, Object>
-        Manifest m = Manifest.fromMap(rawManifest)
+        def rawManifest = new JsonSlurper().parseText('''\
+        {
+          "commerceSuiteVersion": "2005",
+          "useConfig": {
+            "properties": [
+              {
+                "location": "dummy.properties",
+                "persona": "development"
+              },
+              {
+                "location": "dummy.properties",
+                "persona": "invalid"
+              },
+              {
+                "location": "dummy.properties",
+                "aspect": "backoffice"
+              },
+              {
+                "location": "dummy.properties",
+                "aspect": "invalid"
+              }
+            ]
+          }
+        }
+        
+        ''') as Map<String, Object>
+        Manifest manifest = Manifest.fromMap(rawManifest)
         testProjectDir.newFile("dummy.properties")
 
         when:
-        List<Error> errors = validator.validate(m)
+        def errors = validator.validate(manifest)
 
         then:
         errors.size() == 2
@@ -95,8 +165,23 @@ class UseConfigValidatorSpec extends Specification {
 
     def "properties must by a valid Java properties file"() {
         given:
-        def rawManifest = new JsonSlurper().parse(this.getClass().getResource('/validator/useconfig-properties-encoding-manifest.json')) as Map<String, Object>
-        Manifest m = Manifest.fromMap(rawManifest)
+        def rawManifest = new JsonSlurper().parseText('''\
+        {
+          "commerceSuiteVersion": "2005",
+          "useConfig": {
+            "properties": [
+              {
+                "location": "non-latin1.properties",
+              },
+              {
+                "location": "latin1.properties"
+              }
+            ]
+          }
+        }
+        ''') as Map<String, Object>
+        Manifest manifest = Manifest.fromMap(rawManifest)
+
         testProjectDir.newFile("non-latin1.properties").text = '''\
         non.latin1=fööbaß$\\{}
         foo=bar
@@ -107,7 +192,7 @@ class UseConfigValidatorSpec extends Specification {
         '''.stripIndent()
 
         when:
-        List<Error> errors = validator.validate(m)
+        def errors = validator.validate(manifest)
 
         then:
         errors.size() == 1
@@ -116,14 +201,23 @@ class UseConfigValidatorSpec extends Specification {
 
     def "solr customization must have required folder structure"() {
         given:
-        def rawManifest = new JsonSlurper().parse(this.getClass().getResource('/validator/useconfig-solr-manifest.json')) as Map<String, Object>
-        Manifest m = Manifest.fromMap(rawManifest)
+        def rawManifest = new JsonSlurper().parseText('''
+        {
+          "commerceSuiteVersion": "2011",
+          "useConfig": {
+            "solr": {
+              "location": "solr"
+            }
+          }
+        }
+        ''') as Map<String, Object>
+        Manifest manifest = Manifest.fromMap(rawManifest)
         testProjectDir.newFolder("solr")
 
         when:
-        List<Error> missingFolder = validator.validate(m)
+        def missingFolder = validator.validate(manifest)
         testProjectDir.newFolder("solr", "server", "solr", "configsets", "default", "conf")
-        List<Error> folderExists = validator.validate(m)
+        def folderExists = validator.validate(manifest)
 
         then:
         missingFolder.size() == 1